starting to add tests
